cleanup and lint default features
