+++ .github/workflows/build-test.yaml
@@ -28,6 +28,7 @@ jobs:
           -p polars-io \
           -p polars-lazy \
           -- -D warnings
+          cargo clippy
       - name: Feature test
         run: |
           cd polars && cargo hack check --each-feature --no-dev-deps
