+++ polars/polars-core/src/frame/groupby/mod.rs
@@ -1,10 +1,4 @@
-use std::hash::{BuildHasher, Hash, Hasher};
-use std::{
-    fmt::{Debug, Formatter},
-    ops::Add,
-};
-
-use crate::chunked_array::{builder::PrimitiveChunkedBuilder, float::IntegerDecode};
+use crate::chunked_array::builder::PrimitiveChunkedBuilder;
 use crate::frame::select::Selection;
 use crate::prelude::*;
 use crate::utils::{accumulate_dataframes_vertical, split_ca, split_df, NoNull};
@@ -16,8 +10,9 @@ use crate::POOL;
 use ahash::RandomState;
 use hashbrown::{hash_map::RawEntryMut, HashMap};
 use itertools::Itertools;
-use num::{Num, NumCast, Zero};
 use rayon::prelude::*;
+use std::fmt::Debug;
+use std::hash::{BuildHasher, Hash, Hasher};
 
 pub mod aggregations;
 #[cfg(feature = "pivot")]
@@ -486,115 +481,6 @@ impl IntoGroupTuples for ListChunked {}
 #[cfg(feature = "object")]
 impl<T> IntoGroupTuples for ObjectChunked<T> {}
 
-/// Utility enum used for grouping on multiple columns
-#[derive(Copy, Clone, Hash, Eq, PartialEq)]
-pub(crate) enum Groupable<'a> {
-    Boolean(bool),
-    Utf8(&'a str),
-    UInt8(u8),
-    UInt16(u16),
-    UInt32(u32),
-    UInt64(u64),
-    Int8(i8),
-    Int16(i16),
-    Int32(i32),
-    Int64(i64),
-    // mantissa, exponent, sign.
-    Float32(u64, i16, i8),
-    Float64(u64, i16, i8),
-}
-
-impl<'a> Debug for Groupable<'a> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        use Groupable::*;
-        match self {
-            Boolean(v) => write!(f, "{}", v),
-            Utf8(v) => write!(f, "{}", v),
-            UInt8(v) => write!(f, "{}", v),
-            UInt16(v) => write!(f, "{}", v),
-            UInt32(v) => write!(f, "{}", v),
-            UInt64(v) => write!(f, "{}", v),
-            Int8(v) => write!(f, "{}", v),
-            Int16(v) => write!(f, "{}", v),
-            Int32(v) => write!(f, "{}", v),
-            Int64(v) => write!(f, "{}", v),
-            Float32(m, e, s) => write!(f, "float32 mantissa: {} exponent: {} sign: {}", m, e, s),
-            Float64(m, e, s) => write!(f, "float64 mantissa: {} exponent: {} sign: {}", m, e, s),
-        }
-    }
-}
-
-impl From<f64> for Groupable<'_> {
-    fn from(v: f64) -> Self {
-        let (m, e, s) = v.integer_decode();
-        Groupable::Float64(m, e, s)
-    }
-}
-impl From<f32> for Groupable<'_> {
-    fn from(v: f32) -> Self {
-        let (m, e, s) = v.integer_decode();
-        Groupable::Float32(m, e, s)
-    }
-}
-
-fn float_to_groupable_iter<'a, T>(
-    ca: &'a ChunkedArray<T>,
-) -> Box<dyn Iterator<Item = Option<Groupable>> + 'a + Send>
-where
-    T: PolarsNumericType,
-    T::Native: Into<Groupable<'a>>,
-{
-    let iter = ca.into_iter().map(|opt_v| opt_v.map(|v| v.into()));
-    Box::new(iter)
-}
-
-impl<'b> (dyn SeriesTrait + 'b) {
-    pub(crate) fn as_groupable_iter<'a>(
-        &'a self,
-    ) -> Result<Box<dyn Iterator<Item = Option<Groupable>> + 'a + Send>> {
-        macro_rules! as_groupable_iter {
-            ($ca:expr, $variant:ident ) => {{
-                let bx = Box::new($ca.into_iter().map(|opt_b| opt_b.map(Groupable::$variant)));
-                Ok(bx)
-            }};
-        }
-
-        match self.dtype() {
-            DataType::Boolean => as_groupable_iter!(self.bool().unwrap(), Boolean),
-            DataType::UInt8 => as_groupable_iter!(self.u8().unwrap(), UInt8),
-            DataType::UInt16 => as_groupable_iter!(self.u16().unwrap(), UInt16),
-            DataType::UInt32 => as_groupable_iter!(self.u32().unwrap(), UInt32),
-            DataType::UInt64 => as_groupable_iter!(self.u64().unwrap(), UInt64),
-            DataType::Int8 => as_groupable_iter!(self.i8().unwrap(), Int8),
-            DataType::Int16 => as_groupable_iter!(self.i16().unwrap(), Int16),
-            DataType::Int32 => as_groupable_iter!(self.i32().unwrap(), Int32),
-            DataType::Int64 => as_groupable_iter!(self.i64().unwrap(), Int64),
-            DataType::Date32 => {
-                as_groupable_iter!(self.date32().unwrap(), Int32)
-            }
-            DataType::Date64 => {
-                as_groupable_iter!(self.date64().unwrap(), Int64)
-            }
-            DataType::Time64(TimeUnit::Nanosecond) => {
-                as_groupable_iter!(self.time64_nanosecond().unwrap(), Int64)
-            }
-            DataType::Duration(TimeUnit::Nanosecond) => {
-                as_groupable_iter!(self.duration_nanosecond().unwrap(), Int64)
-            }
-            DataType::Duration(TimeUnit::Millisecond) => {
-                as_groupable_iter!(self.duration_millisecond().unwrap(), Int64)
-            }
-            DataType::Utf8 => as_groupable_iter!(self.utf8().unwrap(), Utf8),
-            DataType::Float32 => Ok(float_to_groupable_iter(self.f32().unwrap())),
-            DataType::Float64 => Ok(float_to_groupable_iter(self.f64().unwrap())),
-            DataType::Categorical => as_groupable_iter!(self.categorical().unwrap(), UInt32),
-            dt => Err(PolarsError::Other(
-                format!("Column with dtype {:?} is not groupable", dt).into(),
-            )),
-        }
-    }
-}
-
 impl DataFrame {
     pub fn groupby_with_series(&self, by: Vec<Series>, multithreaded: bool) -> Result<GroupBy> {
         if by.is_empty() || by[0].len() != self.height() {
@@ -722,6 +608,7 @@ pub struct GroupBy<'df, 'selection_str> {
 }
 
 impl<'df, 'selection_str> GroupBy<'df, 'selection_str> {
+    #[cfg(feature = "downsample")]
     fn new(
         df: &'df DataFrame,
         by: Vec<Series>,
