+++ polars/polars-core/src/frame/groupby/pivot.rs
@@ -1,4 +1,121 @@
-use super::*;
+use super::GroupBy;
+use crate::chunked_array::float::IntegerDecode;
+use crate::prelude::*;
+use hashbrown::HashMap;
+use itertools::Itertools;
+use num::{Num, NumCast, Zero};
+use std::collections::hash_map::RandomState;
+use std::fmt::{Debug, Formatter};
+use std::ops::Add;
+
+/// Utility enum used for grouping on multiple columns
+#[derive(Copy, Clone, Hash, Eq, PartialEq)]
+pub(crate) enum Groupable<'a> {
+    Boolean(bool),
+    Utf8(&'a str),
+    UInt8(u8),
+    UInt16(u16),
+    UInt32(u32),
+    UInt64(u64),
+    Int8(i8),
+    Int16(i16),
+    Int32(i32),
+    Int64(i64),
+    // mantissa, exponent, sign.
+    Float32(u64, i16, i8),
+    Float64(u64, i16, i8),
+}
+
+impl<'a> Debug for Groupable<'a> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        use Groupable::*;
+        match self {
+            Boolean(v) => write!(f, "{}", v),
+            Utf8(v) => write!(f, "{}", v),
+            UInt8(v) => write!(f, "{}", v),
+            UInt16(v) => write!(f, "{}", v),
+            UInt32(v) => write!(f, "{}", v),
+            UInt64(v) => write!(f, "{}", v),
+            Int8(v) => write!(f, "{}", v),
+            Int16(v) => write!(f, "{}", v),
+            Int32(v) => write!(f, "{}", v),
+            Int64(v) => write!(f, "{}", v),
+            Float32(m, e, s) => write!(f, "float32 mantissa: {} exponent: {} sign: {}", m, e, s),
+            Float64(m, e, s) => write!(f, "float64 mantissa: {} exponent: {} sign: {}", m, e, s),
+        }
+    }
+}
+
+impl From<f64> for Groupable<'_> {
+    fn from(v: f64) -> Self {
+        let (m, e, s) = v.integer_decode();
+        Groupable::Float64(m, e, s)
+    }
+}
+impl From<f32> for Groupable<'_> {
+    fn from(v: f32) -> Self {
+        let (m, e, s) = v.integer_decode();
+        Groupable::Float32(m, e, s)
+    }
+}
+
+fn float_to_groupable_iter<'a, T>(
+    ca: &'a ChunkedArray<T>,
+) -> Box<dyn Iterator<Item = Option<Groupable>> + 'a + Send>
+where
+    T: PolarsNumericType,
+    T::Native: Into<Groupable<'a>>,
+{
+    let iter = ca.into_iter().map(|opt_v| opt_v.map(|v| v.into()));
+    Box::new(iter)
+}
+
+impl<'b> (dyn SeriesTrait + 'b) {
+    pub(crate) fn as_groupable_iter<'a>(
+        &'a self,
+    ) -> Result<Box<dyn Iterator<Item = Option<Groupable>> + 'a + Send>> {
+        macro_rules! as_groupable_iter {
+            ($ca:expr, $variant:ident ) => {{
+                let bx = Box::new($ca.into_iter().map(|opt_b| opt_b.map(Groupable::$variant)));
+                Ok(bx)
+            }};
+        }
+
+        match self.dtype() {
+            DataType::Boolean => as_groupable_iter!(self.bool().unwrap(), Boolean),
+            DataType::UInt8 => as_groupable_iter!(self.u8().unwrap(), UInt8),
+            DataType::UInt16 => as_groupable_iter!(self.u16().unwrap(), UInt16),
+            DataType::UInt32 => as_groupable_iter!(self.u32().unwrap(), UInt32),
+            DataType::UInt64 => as_groupable_iter!(self.u64().unwrap(), UInt64),
+            DataType::Int8 => as_groupable_iter!(self.i8().unwrap(), Int8),
+            DataType::Int16 => as_groupable_iter!(self.i16().unwrap(), Int16),
+            DataType::Int32 => as_groupable_iter!(self.i32().unwrap(), Int32),
+            DataType::Int64 => as_groupable_iter!(self.i64().unwrap(), Int64),
+            DataType::Date32 => {
+                as_groupable_iter!(self.date32().unwrap(), Int32)
+            }
+            DataType::Date64 => {
+                as_groupable_iter!(self.date64().unwrap(), Int64)
+            }
+            DataType::Time64(TimeUnit::Nanosecond) => {
+                as_groupable_iter!(self.time64_nanosecond().unwrap(), Int64)
+            }
+            DataType::Duration(TimeUnit::Nanosecond) => {
+                as_groupable_iter!(self.duration_nanosecond().unwrap(), Int64)
+            }
+            DataType::Duration(TimeUnit::Millisecond) => {
+                as_groupable_iter!(self.duration_millisecond().unwrap(), Int64)
+            }
+            DataType::Utf8 => as_groupable_iter!(self.utf8().unwrap(), Utf8),
+            DataType::Float32 => Ok(float_to_groupable_iter(self.f32().unwrap())),
+            DataType::Float64 => Ok(float_to_groupable_iter(self.f64().unwrap())),
+            DataType::Categorical => as_groupable_iter!(self.categorical().unwrap(), UInt32),
+            dt => Err(PolarsError::Other(
+                format!("Column with dtype {:?} is not groupable", dt).into(),
+            )),
+        }
+    }
+}
 
 impl<'df, 'selection_str> GroupBy<'df, 'selection_str> {
     /// Pivot a column of the current `DataFrame` and perform one of the following aggregations:
